Tslint and checkstyle

+++ arm-front/src/app/components/apartment-detail/apartment-detail.component.ts
@@ -1,10 +1,15 @@
 import { Component, OnInit, Input, ViewChild, ElementRef } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Location } from '@angular/common';
+
+import { } from '@types/googlemaps';
+
 import { Apartment } from '../../models/apartment';
 import { ApartmentService } from '../../services/apartment/apartment.service';
 import { GeocodingService } from '../../services/geocoding/geocoding.service';
-import { } from '@types/googlemaps';
+import { AuthService } from '../../services/auth/auth.service';
+import { User } from '../../models/user';
+import { UserService } from '../../services/user/user.service';
 
 @Component({
   selector: 'app-apartment-detail',
@@ -15,31 +20,41 @@ export class ApartmentDetailComponent implements OnInit {
 
   @ViewChild('gmap') gmapElement: any;
   @ViewChild('nameInput') input: ElementRef;
-  @Input() apartment : Apartment;
+  @Input() apartment: Apartment;
   editable = false;
   isNew = false;
+  error: boolean;
   map: google.maps.Map;
   marker: google.maps.Marker;
-  address: string = "";
+  address = '';
+  realtors: User[];
 
   constructor(
     private route: ActivatedRoute,
     private apartmentService: ApartmentService,
     private geocodingService: GeocodingService,
-    private location: Location) { }
+    private location: Location,
+    private auth: AuthService,
+    private userService: UserService) { }
 
-    
   ngOnInit() {
     this.getApartment();
     this.mapInitGeo();
+
+    if (this.auth.roleMatch(['ADMIN'])) {
+      this.userService.getRealtors()
+      .subscribe(realtors => {
+        this.realtors = realtors;
+      });
+    }
   }
 
   modelChanged(newObj) {
     this.mapSetMarker();
   }
 
-  mapSetMarker() : void {
-    var center = {lat: +this.apartment.latitude, lng: +this.apartment.longitude};
+  mapSetMarker(): void {
+    const center = {lat: +this.apartment.latitude, lng: +this.apartment.longitude};
     this.map.setCenter(center);
     if (this.marker) {
       this.marker.setMap(null);
@@ -47,15 +62,15 @@ export class ApartmentDetailComponent implements OnInit {
     this.marker = new google.maps.Marker({position: center, map: this.map});
   }
 
-  mapInitGeo() : void {
+  mapInitGeo(): void {
     this.map = new google.maps.Map(this.gmapElement.nativeElement, {
       center: {lat: -34.397, lng: 150.644},
       zoom: 6
     });
 
-    var parent = this;
+    const parent = this;
     this.map.addListener('click', function(e) {
-      if (parent.editable == true) {
+      if (parent.editable === true || parent.isNew === true) {
         parent.apartment.longitude = e.latLng.lng();
         parent.apartment.latitude = e.latLng.lat();
         parent.mapSetMarker();
@@ -64,10 +79,9 @@ export class ApartmentDetailComponent implements OnInit {
       }
     });
 
-    var parent = this;
     if (navigator.geolocation) {
       navigator.geolocation.getCurrentPosition(function(position) {
-        var pos = {
+        const pos = {
           lat: position.coords.latitude,
           lng: position.coords.longitude
         };
@@ -79,7 +93,7 @@ export class ApartmentDetailComponent implements OnInit {
     }
   }
 
-  getApartment() : void {
+  getApartment(): void {
     const id = +this.route.snapshot.paramMap.get('id');
     if (id) {
       this.apartmentService.getApartment(id)
@@ -95,12 +109,38 @@ export class ApartmentDetailComponent implements OnInit {
 
   saveApartment() {
     this.apartmentService.updateApartment(this.apartment)
-    .subscribe(() => this.goBack())
+    .subscribe(data => {
+      if (data.error) {
+        this.error = data.message;
+      } else {
+        this.goBack();
+      }
+    });
+  }
+
+  setFree(): void {
+    if (this.apartment.isRented) {
+      this.apartmentService.freeApartment(this.apartment)
+      .subscribe(() => alert('Apartment is available now'));
+      this.apartment.isRented = false;
+    }
+  }
+
+  setRented(): void {
+    if (!this.apartment.isRented) {
+      this.apartmentService.rentApartment(this.apartment)
+      .subscribe(() => alert('Apartment is rented now'));
+      this.apartment.isRented = true;
+    }
   }
 
   deleteApartment() {
-    this.apartmentService.deleteApartment(this.apartment)
-    .subscribe(() => this.goBack())
+    if (confirm('Are you sure?')) {
+      this.apartmentService.deleteApartment(this.apartment)
+    .subscribe(() => this.goBack());
+    } else {
+      return;
+    }
   }
 
   cancelEditing() {
@@ -112,15 +152,15 @@ export class ApartmentDetailComponent implements OnInit {
     this.location.back();
   }
 
-  showInputs() : boolean {
-    return this.editable == true || this.isNew == true;
+  showInputs(): boolean {
+    return this.editable === true || this.isNew === true;
   }
 
-  getCoordinates() : void {
-    var parent = this;
+  getCoordinates(): void {
+    const parent = this;
     this.geocodingService.getCoordinates(this.address)
     .subscribe(data => {
-      if (data.status == "OK" && data.results) {
+      if (data.status === 'OK' && data.results) {
         parent.address = data.results[0].formatted_address;
         parent.apartment.longitude = data.results[0].geometry.location.lng;
         parent.apartment.latitude = data.results[0].geometry.location.lat;
@@ -130,5 +170,4 @@ export class ApartmentDetailComponent implements OnInit {
       }
     });
   }
-
 }
