+++ arm-front/src/app/components/user-detail/user-detail.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, Input, ViewChild, ElementRef } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Location } from '@angular/common';
+
 import { User } from '../../models/user';
 import { UserService } from '../../services/user/user.service';
 
@@ -11,10 +12,15 @@ import { UserService } from '../../services/user/user.service';
 })
 export class UserDetailComponent implements OnInit {
 
-  @Input() user : User;
+  @Input() user: User;
 
   editable = false;
   isNew = false;
+
+  error: string;
+
+  newPassword: string;
+  passwordError: string;
 
   constructor(private route: ActivatedRoute,
     private userService: UserService,
@@ -24,7 +30,7 @@ export class UserDetailComponent implements OnInit {
     this.getUser();
   }
 
-  getUser() : void {
+  getUser(): void {
     const id = +this.route.snapshot.paramMap.get('id');
     if (id) {
       this.userService.getUser(id)
@@ -37,14 +43,41 @@ export class UserDetailComponent implements OnInit {
     }
   }
 
+  isValid(): boolean {
+    if (!this.user.name) {
+      return false;
+    }
+    if (!this.user.login) {
+      return false;
+    }
+    if (!this.user.roleType) {
+      return false;
+    }
+    return true;
+  }
+
   saveUser() {
-    this.userService.updateUser(this.user)
-    .subscribe(() => this.goBack())
+    if (this.isValid()) {
+      this.userService.updateUser(this.user)
+      .subscribe(data => {
+        if (data) {
+          this.error = data.message;
+        } else {
+          this.goBack();
+        }
+      });
+    } else {
+      this.error = 'Fill all fields';
+    }
   }
 
   deleteUser() {
-    this.userService.deleteUser(this.user)
-    .subscribe(() => this.goBack())
+    if (confirm('Are you sure?')) {
+      this.userService.deleteUser(this.user)
+    .subscribe(() => this.goBack());
+    } else {
+      return;
+    }
   }
 
   cancelEditing() {
@@ -56,8 +89,23 @@ export class UserDetailComponent implements OnInit {
     this.location.back();
   }
 
-  showInputs() : boolean {
-    return this.editable == true || this.isNew == true;
+  showInputs(): boolean {
+    return this.editable === true || this.isNew === true;
   }
 
+  changePassword(): void {
+    const passdata = {
+      id : this.user.id,
+      password : this.newPassword
+    };
+    this.userService.changePassword(passdata)
+      .subscribe(data => {
+        if (data) {
+          this.passwordError = data.message;
+        } else {
+          alert('Password successfully changed');
+          this.newPassword = '';
+        }
+      });
+  }
 }
